requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
pymupdf>=1.23.8
openai>=1.6.1
pinecone>=5.0.0
python-multipart>=0.0.6
requests>=2.31.0
zeroentropy>=0.1.0a6
supabase>=2.0.0
postgrest>=0.10.0
